almost get the vectorized sampler working, but have off by one errors and strange behavior
